+++ src/Aks/Aks/Commands/SetAzureRmAks.cs
@@ -416,6 +416,10 @@ namespace Microsoft.Azure.Commands.Aks
                             cluster.Sku = new ManagedClusterSKU(name: "Basic", tier: "Free");
                         }
                     }
+                    if (this.IsParameterBound(c => c.AadProfile))
+                    {
+                        cluster.AadProfile = AadProfile;
+                    }
                     SetIdentity(cluster);
 
                     var kubeCluster = this.CreateOrUpdate(ResourceGroupName, Name, cluster);
@@ -428,6 +432,7 @@ namespace Microsoft.Azure.Commands.Aks
                     {
                         cluster.DisableLocalAccounts = DisableLocalAccount;
                     }
+                    
                     WriteObject(PSMapper.Instance.Map<PSKubernetesCluster>(kubeCluster));
                 });
             }