+++ src/Cdn/exports/Remove-AzFrontDoorCdnSecurityPolicy.ps1
@@ -46,7 +46,7 @@ INPUTOBJECT <ICdnIdentity>: Identity Parameter
   [SecurityPolicyName <String>]: Name of the security policy under the profile.
   [SubscriptionId <String>]: Azure Subscription ID.
 .Link
-https://docs.microsoft.com/powershell/module/az.cdn/remove-azfrontdoorcdnsecuritypolicy
+https://learn.microsoft.com/powershell/module/az.cdn/remove-azfrontdoorcdnsecuritypolicy
 #>
 function Remove-AzFrontDoorCdnSecurityPolicy {
 [OutputType([System.Boolean])]