+++ src/ContainerInstance/exports/Get-AzContainerInstanceContainerGroupOutboundNetworkDependencyEndpoint.ps1
@@ -41,7 +41,7 @@ INPUTOBJECT <IContainerInstanceIdentity>: Identity Parameter
   [ResourceGroupName <String>]: The name of the resource group.
   [SubscriptionId <String>]: Subscription credentials which uniquely identify Microsoft Azure subscription. The subscription ID forms part of the URI for every service call.
 .Link
-https://docs.microsoft.com/powershell/module/az.containerinstance/get-azcontainerinstancecontainergroupoutboundnetworkdependencyendpoint
+https://learn.microsoft.com/powershell/module/az.containerinstance/get-azcontainerinstancecontainergroupoutboundnetworkdependencyendpoint
 #>
 function Get-AzContainerInstanceContainerGroupOutboundNetworkDependencyEndpoint {
 [Alias('Get-AzContainerGroupOutboundNetworkDependencyEndpoint')]