+++ src/Aks/Aks/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added parameter `-AadProfile` for `New-AzAksCluster` and `Set-AzAksCluster`
+* Added parameter `-NodeHostGroupID` for `New-AzAksCluster` and parameter `-HostGroupID` for `New-AzAksNodePool`
+
+## Version 5.2.0
 * Added parameter `-EnableEncryptionAtHost` for `New-AzAksCluster` and `New-AzAksNodePool`
 * Added parameter `-EnableUltraSSD` for `New-AzAksCluster` and `New-AzAksNodePool`
 * Added parameter `-NodeKubeletConfig` for `New-AzAksCluster`, `-KubeletConfig` for `New-AzAksNodePool`
@@ -29,6 +33,7 @@
 * Added parameter `-AutoScalerProfile` for `New-AzAksCluster` and `Set-AzAksCluster`
 * Added parameter `-GpuInstanceProfile` for `New-AzAksCluster` and `New-AzAksNodePool`
 * Added parameter `-EnableUptimeSLA` for `New-AzAksCluster` and `Set-AzAksCluster`
+* Added parameter `-EdgeZone` for `New-AzAksCluster`
 
 ## Version 5.1.0
 * Bumped API version to 2022-09-01