+++ src/Resources/Resources/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Fixed an issue where running deployment cmdlets with `-WhatIf` throws exception when formatting results with nested array changes
 
 ## Version 6.5.1
 * Fixed issue introduced in previous fix for `Set-AzPolicySetDefinition` InternalServerError when the initiative is too large [#20238], which will remove space in value.